Add optional support for multiple pings per probe
By adding the `packet_count` option to the module, multiple pings
(between 2 and 10) will be sent to the target. In this mode
`probe_icmp_duration_seconds` becomes the _average_ of the RTT for all
the pings, `probe_icmp_duration_seconds_max` is the maximum observed
value, `probe_icmp_duration_seconds_min` the mininum and
`probe_icmp_duration_seconds_stddev` the standard deviation. This also
adds `probe_icmp_sent_packets` to record the number of sent packets and
`probe_icmp_received_packets` the number of received packets. Probe
success is reported _only_ if all the sent packets are returned. _Any_
packet loss is reported as a failure.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>