Add ip_filter support to block certain IP addresses
IPFilter is used to filter network traffic so that probes are limited in
the destination addresses they can probe.

The configuration looks like:
      ip_filter:
               default: allow|deny
               allowed:
                 [... list of networks in CIDR format ...]
               denied:
                 [... list of networks in CIDR format ...]

This can be specified at the top level or inside each module definition.

The longest prefix match takes precedence:

    ip_filter:
      default: deny
      allowed:
        - 10.0.0.0/8
      denied:
        - 10.0.1.0/24

Only traffic to 10.0.0.0/8 is allowed, except the 10.0.1.0/24 subnet
which is blocked.

The default key is only mandatory when both allowed and denied
are defined. Otherwise, it is implicitly deny or allow.

Signed-off-by: Adrian Serrano <adrisr83@gmail.com>